# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SaaS FinOps Analytics Platform API
  version: 1.0.0
  description: |
    The SaaS FinOps Analytics Platform provides multi-tenant cloud cost management,
    AI-powered database analytics, and subscription billing capabilities.

    ## Authentication

    The API supports two authentication methods:

    ### JWT Bearer Token
    Obtain a token via `POST /auth/login`. Include it in the `Authorization` header:
    ```
    Authorization: Bearer <access_token>
    ```
    Access tokens expire after **15 minutes**. Use `POST /auth/refresh` to obtain a new one.

    ### API Key
    Generate an API key via `POST /auth/api-keys`. Include it in the `X-API-Key` header:
    ```
    X-API-Key: <your_api_key>
    ```

    ## Rate Limiting

    Rate limits are enforced per user per minute based on subscription tier:

    | Plan       | Requests/min |
    |------------|-------------|
    | Free       | 100         |
    | Base       | 500         |
    | Pro        | 2,000       |
    | Enterprise | 10,000      |

    When the limit is exceeded, the API returns `429 Too Many Requests` with a `Retry-After` header.

  contact:
    name: Platform Support
    email: support@finops-platform.example.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development (API Gateway)
  - url: http://localhost:8081
    description: Auth Service (direct)
  - url: http://localhost:8082
    description: Billing Service (direct)
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from POST /auth/login. Expires in 15 minutes.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key generated via POST /auth/api-keys.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "invalid credentials"
        message:
          type: string
          example: "The provided email or password is incorrect"
      required: [error]
...
    Recommendation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [idle_resource, oversized_resource, unattached_storage]
          example: "idle_resource"
        resource_id:
          type: string
          example: "i-1234567890abcdef0"
        description:
          type: string
          example: "EC2 instance with 0% CPU utilization for 7 days"
        potential_monthly_savings:
          type: number
          example: 150.00
        cloud_account_id:
          type: string
          format: uuid
    RecommendationsResponse:
      type: object
      properties:
        recommendations:
          type: array
          items:
            $ref: '#/components/schemas/Recommendation'
        total_potential_savings:
          type: number
          example: 2340.00
...
paths:
...
  /finops/recommendations:
    get:
      tags: [finops]
      summary: Get cost optimization recommendations
      description: |
        Returns cost optimization recommendations sorted by potential savings.
        Recommendations are updated daily at 2:00 AM UTC.

        **Recommendation types:**
        - `idle_resource`: Zero usage for 7+ consecutive days
        - `oversized_resource`: Average utilization <20% over 30 days
        - `unattached_storage`: Storage volumes incurring costs with no attachment
      operationId: getRecommendations
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      responses:
        '200':
          description: Optimization recommendations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...